+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add support for stepped ranges in Elixir 1.12
 * Add `A.Enum.concat/1`
+* Add `A.Vector.zip_with/3`
 
 ### Bug fixes